refactor(frontend) parseProcessInformation (#819)
